--- gsuite/util.go
@@ -1,17 +1,0 @@
-package gsuite
-
-import "github.com/hashicorp/terraform/helper/schema"
-
-func mergeSchemas(a, b map[string]*schema.Schema) map[string]*schema.Schema {
-	merged := make(map[string]*schema.Schema)
-
-	for k, v := range a {
-		merged[k] = v
-	}
-
-	for k, v := range b {
-		merged[k] = v
-	}
-
-	return merged
-}
