+++ gsuite/utils.go
@@ -40,3 +40,17 @@ func retryTime(retryFunc func() error, minutes int) error {
 		return resource.NonRetryableError(err)
 	})
 }
+
+func mergeSchemas(a, b map[string]*schema.Schema) map[string]*schema.Schema {
+	merged := make(map[string]*schema.Schema)
+
+	for k, v := range a {
+		merged[k] = v
+	}
+
+	for k, v := range b {
+		merged[k] = v
+	}
+
+	return merged
+}
